refactor(widgets): extract AppGridWidget and LocaleSwitcherWidget
- AppGridWidget: app grid with search and shortcuts; MenuWidget composes
backdrop + panel + grid
- LocaleSwitcherWidget: locale trigger and dropdown; TopBarRightWidget
composes switcher + icons + account menu
